+++ Xcode/Bluetooth.xcodeproj/xcshareddata/xcbaselines/6EE84DC01CAF5C7C00A40C4D.xcbaseline/E272A968-C86F-4169-956E-E6220948798B.plist
@@ -17,6 +17,89 @@
 				</dict>
 			</dict>
 		</dict>
+		<key>UUIDTests</key>
+		<dict>
+			<key>testPerfomanceStringCoreBluetooth()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.075649</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerfomanceStringParseCoreBluetooth()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.11503</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceData()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.072725</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceDataCoreBluetooth()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.11413</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceDataParse()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.070214</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceDataParseCoreBluetooth()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.073602</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceString()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.088214</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+			<key>testPerformanceStringParse()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.16358</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>Apr 4, 2018 at 10:19:56 PM</string>
+				</dict>
+			</dict>
+		</dict>
 	</dict>
 </dict>
 </plist>
