+++ Xcode/Bluetooth.xcodeproj/xcshareddata/xcschemes/Bluetooth-macOS.xcscheme
@@ -27,7 +27,17 @@
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       codeCoverageEnabled = "YES"
+      onlyGenerateCoverageForSpecifiedTargets = "YES"
       shouldUseLaunchSchemeArgsEnv = "YES">
+      <CodeCoverageTargets>
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "6EE84DB61CAF5C7C00A40C4D"
+            BuildableName = "Bluetooth.framework"
+            BlueprintName = "Bluetooth-macOS"
+            ReferencedContainer = "container:Bluetooth.xcodeproj">
+         </BuildableReference>
+      </CodeCoverageTargets>
       <Testables>
          <TestableReference
             skipped = "NO">
